requirements.txt
Flask==3.0.3
opencv-python==4.10.0.84
Pillow==10.4.0
numpy==2.0.2
pytesseract==0.3.13
gunicorn==21.2.0